option to keep terminator

+++ terminator/install.sh
@@ -1,4 +1,5 @@
 #!/bin/sh
 [ "$(uname -s)" = "Darwin" ] && exit 0
+[ "$KEEP_TERMINATOR" = "true" ] && exit 0
 mkdir -p ~/.config/terminator/
 ln -sf "$ZSH"/terminator/config ~/.config/terminator/config
